Update the ranking criteria to specify Source rules
Currently there are no rules on the source and it has caused debate and issues as to whether or not remixes for example should have it's original material/album as the source.

These rules set the record straight for the future and makes it extremely clear what the rules on the source are.